requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
redis>=5.0.4
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# Google Gemini
google-genai>=1.0.0

# Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# Retrieval (optional — Pinecone disabled if PINECONE_API_KEY not set)
pinecone>=3.0.0

# EDA Pipeline
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0

# MCP Server
mcp>=1.0.0
httpx>=0.27.0
